requirements.txt
black>=21.5b2
deepface>=0.0.75
flake8>=3.9.0
isort>=5.9.0
jinja2>=3.0.0
matplotlib>=3.3.0
mypy>=0.910
# Core dependencies
numpy>=1.19.0
opencv-python>=4.5.0
pandas>=1.3.0
pdfkit>=1.0.0
pillow>=8.0.0
plotly>=5.3.0
pre-commit>=2.15.0

# Development dependencies
pytest>=6.2.5
pytest-cov>=2.12.0
pyyaml>=5.4.1
scikit-learn>=0.24.0
scipy>=1.7.0
seaborn>=0.11.0
torch>=1.9.0
torchvision>=0.10.0
